refactor: Extract State and ResourcesState from into separate components
- Create dedicated DeploymentState.tsx with state logic,
types and styling
- Create dedicated DeploymentResourcesState.tsx with resources state
logic, types, and styling

This refactoring improves code organization and makes
the deployment state components more reusable across the application.
